Decouple script parsing from the revision grid
