+++ GitUI/Script/ScriptOptionsParser.cs
@@ -75,7 +75,7 @@ namespace GitUI.Script
             return option.StartsWith("s");
         }
 
-        public static (string arguments, bool abort) Parse(string arguments, IGitModule module, IWin32Window owner, RevisionGridControl revisionGrid)
+        public static (string arguments, bool abort) Parse(string arguments, IGitModule module, IWin32Window owner, IScriptHostControl scriptHostControl)
         {
             GitRevision selectedRevision = null;
             GitRevision currentRevision = null;
@@ -106,7 +106,7 @@ namespace GitUI.Script
 
                 if (currentRevision == null && option.StartsWith("c"))
                 {
-                    currentRevision = GetCurrentRevision(module, revisionGrid, currentTags, currentLocalBranches, currentRemoteBranches, currentBranches);
+                    currentRevision = GetCurrentRevision(module, scriptHostControl, currentTags, currentLocalBranches, currentRemoteBranches, currentBranches);
                     if (currentRevision == null)
                     {
                         return (arguments: null, abort: true);
@@ -122,21 +122,21 @@ namespace GitUI.Script
                         if (string.IsNullOrEmpty(currentRemote))
                         {
                             currentRemote = module.GetSetting(string.Format(SettingKeyString.BranchRemote,
-                                AskToSpecify(currentLocalBranches, revisionGrid)));
+                                AskToSpecify(currentLocalBranches, scriptHostControl)));
                         }
                     }
                 }
-                else if (selectedRevision == null && revisionGrid != null && DependsOnSelectedRevision(option))
+                else if (selectedRevision == null && scriptHostControl != null && DependsOnSelectedRevision(option))
                 {
-                    allSelectedRevisions = revisionGrid.GetSelectedRevisions();
-                    selectedRevision = CalculateSelectedRevision(revisionGrid, selectedRemoteBranches, selectedRemotes, selectedLocalBranches, selectedBranches, selectedTags);
+                    allSelectedRevisions = scriptHostControl.GetSelectedRevisions();
+                    selectedRevision = CalculateSelectedRevision(scriptHostControl, selectedRemoteBranches, selectedRemotes, selectedLocalBranches, selectedBranches, selectedTags);
                     if (selectedRevision == null)
                     {
                         return (arguments: null, abort: true);
                     }
                 }
 
-                arguments = ParseScriptArguments(arguments, option, owner, revisionGrid, module, allSelectedRevisions, selectedTags, selectedBranches, selectedLocalBranches, selectedRemoteBranches, selectedRemotes, selectedRevision, currentTags, currentBranches, currentLocalBranches, currentRemoteBranches, currentRevision, currentRemote);
+                arguments = ParseScriptArguments(arguments, option, owner, scriptHostControl, module, allSelectedRevisions, selectedTags, selectedBranches, selectedLocalBranches, selectedRemoteBranches, selectedRemotes, selectedRevision, currentTags, currentBranches, currentLocalBranches, currentRemoteBranches, currentRevision, currentRemote);
                 if (arguments == null)
                 {
                     return (arguments: null, abort: true);
@@ -146,33 +146,33 @@ namespace GitUI.Script
             return (arguments, abort: false);
         }
 
-        private static string AskToSpecify(IEnumerable<IGitRef> options, RevisionGridControl revisionGrid)
+        private static string AskToSpecify(IEnumerable<IGitRef> options, IScriptHostControl scriptHostControl)
         {
             using (var f = new FormQuickGitRefSelector())
             {
-                f.Location = revisionGrid?.GetQuickItemSelectorLocation() ?? new System.Drawing.Point();
+                f.Location = scriptHostControl?.GetQuickItemSelectorLocation() ?? new System.Drawing.Point();
                 f.Init(FormQuickGitRefSelector.Action.Select, options.ToList());
                 f.ShowDialog();
                 return f.SelectedRef.Name;
             }
         }
 
-        private static string AskToSpecify(IEnumerable<string> options, RevisionGridControl revisionGrid)
+        private static string AskToSpecify(IEnumerable<string> options, IScriptHostControl scriptHostControl)
         {
             using (var f = new FormQuickStringSelector())
             {
-                f.Location = revisionGrid?.GetQuickItemSelectorLocation() ?? new System.Drawing.Point();
+                f.Location = scriptHostControl?.GetQuickItemSelectorLocation() ?? new System.Drawing.Point();
                 f.Init(options.ToList());
                 f.ShowDialog();
                 return f.SelectedString;
             }
         }
 
-        private static GitRevision CalculateSelectedRevision(RevisionGridControl revisionGrid, List<IGitRef> selectedRemoteBranches,
+        private static GitRevision CalculateSelectedRevision(IScriptHostControl scriptHostControl, List<IGitRef> selectedRemoteBranches,
             List<string> selectedRemotes, List<IGitRef> selectedLocalBranches,
             List<IGitRef> selectedBranches, List<IGitRef> selectedTags)
         {
-            GitRevision selectedRevision = revisionGrid.LatestSelectedRevision;
+            GitRevision selectedRevision = scriptHostControl.GetLatestSelectedRevision();
             if (selectedRevision == null)
             {
                 return null;
@@ -207,12 +207,12 @@ namespace GitUI.Script
 
         [CanBeNull]
         private static GitRevision GetCurrentRevision(
-            IGitModule module, [CanBeNull] RevisionGridControl revisionGrid, List<IGitRef> currentTags, List<IGitRef> currentLocalBranches,
+            IGitModule module, [CanBeNull] IScriptHostControl scriptHostControl, List<IGitRef> currentTags, List<IGitRef> currentLocalBranches,
             List<IGitRef> currentRemoteBranches, List<IGitRef> currentBranches)
         {
             GitRevision currentRevision;
             IEnumerable<IGitRef> refs;
-            if (revisionGrid == null)
+            if (scriptHostControl == null)
             {
                 var currentRevisionGuid = module.GetCurrentCheckout();
                 currentRevision = currentRevisionGuid == null ? null : new GitRevision(currentRevisionGuid);
@@ -220,7 +220,7 @@ namespace GitUI.Script
             }
             else
             {
-                currentRevision = revisionGrid.GetCurrentRevision();
+                currentRevision = scriptHostControl.GetCurrentRevision();
                 refs = currentRevision.Refs;
             }
 
@@ -258,7 +258,7 @@ namespace GitUI.Script
             return "";
         }
 
-        private static string ParseScriptArguments(string arguments, string option, IWin32Window owner, RevisionGridControl revisionGrid, IGitModule module, IReadOnlyList<GitRevision> allSelectedRevisions, in IList<IGitRef> selectedTags, in IList<IGitRef> selectedBranches, in IList<IGitRef> selectedLocalBranches, in IList<IGitRef> selectedRemoteBranches, in IList<string> selectedRemotes, GitRevision selectedRevision, in IList<IGitRef> currentTags, in IList<IGitRef> currentBranches, in IList<IGitRef> currentLocalBranches, in IList<IGitRef> currentRemoteBranches, GitRevision currentRevision, string currentRemote)
+        private static string ParseScriptArguments(string arguments, string option, IWin32Window owner, IScriptHostControl scriptHostControl, IGitModule module, IReadOnlyList<GitRevision> allSelectedRevisions, in IList<IGitRef> selectedTags, in IList<IGitRef> selectedBranches, in IList<IGitRef> selectedLocalBranches, in IList<IGitRef> selectedRemoteBranches, in IList<string> selectedRemotes, GitRevision selectedRevision, in IList<IGitRef> currentTags, in IList<IGitRef> currentBranches, in IList<IGitRef> currentLocalBranches, in IList<IGitRef> currentRemoteBranches, GitRevision currentRevision, string currentRemote)
         {
             string newString = null;
             string remote;
@@ -467,27 +467,27 @@ namespace GitUI.Script
 
             return arguments;
 
-            string SelectOneRef(IList<IGitRef> refs) => ScriptOptionsParser.SelectOne(refs, revisionGrid);
-            string SelectOneString(IList<string> strings) => ScriptOptionsParser.SelectOne(strings, revisionGrid);
+            string SelectOneRef(IList<IGitRef> refs) => ScriptOptionsParser.SelectOne(refs, scriptHostControl);
+            string SelectOneString(IList<string> strings) => ScriptOptionsParser.SelectOne(strings, scriptHostControl);
         }
 
-        private static string SelectOne(IList<IGitRef> refs, RevisionGridControl revisionGrid)
+        private static string SelectOne(IList<IGitRef> refs, IScriptHostControl scriptHostControl)
         {
             switch (refs.Count)
             {
                 case 0: return string.Empty;
                 case 1: return refs[0].Name;
-                default: return AskToSpecify(refs, revisionGrid);
+                default: return AskToSpecify(refs, scriptHostControl);
             }
         }
 
-        private static string SelectOne(IList<string> strings, RevisionGridControl revisionGrid)
+        private static string SelectOne(IList<string> strings, IScriptHostControl scriptHostControl)
         {
             switch (strings.Count)
             {
                 case 0: return string.Empty;
                 case 1: return strings[0];
-                default: return AskToSpecify(strings, revisionGrid);
+                default: return AskToSpecify(strings, scriptHostControl);
             }
         }
 
@@ -511,8 +511,8 @@ namespace GitUI.Script
 
         internal readonly struct TestAccessor
         {
-            public string ParseScriptArguments(string arguments, string option, IWin32Window owner, RevisionGridControl revisionGrid, IGitModule module, IReadOnlyList<GitRevision> allSelectedRevisions, List<IGitRef> selectedTags, List<IGitRef> selectedBranches, List<IGitRef> selectedLocalBranches, List<IGitRef> selectedRemoteBranches, List<string> selectedRemotes, GitRevision selectedRevision, List<IGitRef> currentTags, List<IGitRef> currentBranches, List<IGitRef> currentLocalBranches, List<IGitRef> currentRemoteBranches, GitRevision currentRevision, string currentRemote) =>
-                ScriptOptionsParser.ParseScriptArguments(arguments, option, owner, revisionGrid, module, allSelectedRevisions, selectedTags, selectedBranches, selectedLocalBranches, selectedRemoteBranches, selectedRemotes, selectedRevision, currentTags, currentBranches, currentLocalBranches, currentRemoteBranches, currentRevision, currentRemote);
+            public string ParseScriptArguments(string arguments, string option, IWin32Window owner, IScriptHostControl scriptHostControl, IGitModule module, IReadOnlyList<GitRevision> allSelectedRevisions, List<IGitRef> selectedTags, List<IGitRef> selectedBranches, List<IGitRef> selectedLocalBranches, List<IGitRef> selectedRemoteBranches, List<string> selectedRemotes, GitRevision selectedRevision, List<IGitRef> currentTags, List<IGitRef> currentBranches, List<IGitRef> currentLocalBranches, List<IGitRef> currentRemoteBranches, GitRevision currentRevision, string currentRemote) =>
+                ScriptOptionsParser.ParseScriptArguments(arguments, option, owner, scriptHostControl, module, allSelectedRevisions, selectedTags, selectedBranches, selectedLocalBranches, selectedRemoteBranches, selectedRemotes, selectedRevision, currentTags, currentBranches, currentLocalBranches, currentRemoteBranches, currentRevision, currentRemote);
         }
     }
 }
